+++ packages/fes-plugin-layout/src/index.js
@@ -32,9 +32,6 @@ export default (api) => {
 
         const HAS_LOCALE = api.hasPlugins(['@fesjs/plugin-locale']);
 
-        // 路由信息
-        const routes = await api.getRoutes();
-
         // .fes配置
         const userConfig = {
             title: name,
@@ -42,11 +39,7 @@ export default (api) => {
             ...(api.config.layout || {})
         };
 
-        // 把路由的meta合并到menu配置中
-        const menus = helper.fillMenuByRoute(userConfig.menus, routes);
-        delete userConfig.menus;
-
-        const iconNames = helper.getIconNamesFromMenu(menus);
+        const iconNames = helper.getIconNamesFromMenu(userConfig.menus);
 
         const iconsString = iconNames.map(
             iconName => `import { ${iconName} } from '@fesjs/fes-design/icon'`
@@ -66,7 +59,6 @@ export default (api) => {
                 readFileSync(join(__dirname, 'runtime/index.tpl'), 'utf-8'),
                 {
                     REPLACE_USER_CONFIG: JSON.stringify(userConfig),
-                    MENUS: JSON.stringify(menus),
                     HAS_LOCALE
                 }
             )
