+++ packages/fes-plugin-layout/src/runtime/index.tpl
@@ -1,49 +1,25 @@
-import { ref, defineComponent, inject } from "vue";
-import { plugin, ApplyPluginsType } from "@@/core/coreExports";
-import BaseLayout from "./views/BaseLayout.vue";
+import { ref, defineComponent, computed } from 'vue';
+import { plugin, ApplyPluginsType, getRoutes } from '@@/core/coreExports';
+import BaseLayout from './views/BaseLayout.vue';
+import getRuntimeConfig from './helpers/getRuntimeConfig';
+import fillMenu from './helpers/fillMenu';
 
 const Layout = defineComponent({
     name: 'Layout',
-    setup(){
+    setup() {
         const userConfig = {{{REPLACE_USER_CONFIG}}};
-        const menusRef = ref({{{MENUS}}})
-        const runtimeConfig = plugin.applyPlugins({
-            key: "layout",
-            type: ApplyPluginsType.modify,
-            initialValue: {
-                initialState: inject('initialState')
-            },
-        });
+        const runtimeConfig = getRuntimeConfig();
+        let menusRef = ref(userConfig.menus);
         // 如果运行时配置了menus，则需要处理
-        if (
-            runtimeConfig.menus &&
-            typeof runtimeConfig.menus.request === 'function'
-        ) {
-            const fetchData = async () => {
-                try {
-                    const newMenusData = await runtimeConfig.menus.request(
-                        menusRef.value
-                    );
-                    return newMenusData;
-                } catch (e) {
-                    console.error('[plugin-layout]: runtime config menus.request run error:', e);
-                }
-                return menusRef.value;
-            };
-            if (runtimeConfig.menus.watch) {
-                watch(
-                    () => runtimeConfig.menus.watch,
-                    async () => {
-                        const newMenusData = await fetchData();
-                        menusRef.value = newMenusData;
-                    }
-                );
-            }
-            fetchData().then((newMenusData) => {
-                menusRef.value = newMenusData;
-            });
+        if (runtimeConfig.menus && typeof runtimeConfig.menus === 'function') {
+            menusRef = ref(runtimeConfig.menus(userConfig.menus));
         }
-        const localeShared = plugin.getShared("locale");
+        // 把路由的meta合并到menu配置中
+        const filledMenuRef = computed(() => {
+            return fillMenu(menusRef.value, getRoutes());
+        });
+
+        const localeShared = plugin.getShared('locale');
         return () => {
             const slots = {
                 customHeader: () => {
@@ -56,14 +32,23 @@ const Layout = defineComponent({
                 },
                 locale: () => {
                     if (localeShared) {
-                        return <localeShared.SelectLang></localeShared.SelectLang>;
+                        return (
+                            <localeShared.SelectLang></localeShared.SelectLang>
+                        );
                     }
                     return null;
-                },
+                }
             };
-            return <BaseLayout locale={ localeShared ? true : false } {...userConfig} menus={menusRef.value} v-slots={slots}></BaseLayout>;
+            return (
+                <BaseLayout
+                    locale={localeShared ? true : false}
+                    {...userConfig}
+                    menus={filledMenuRef.value}
+                    v-slots={slots}
+                ></BaseLayout>
+            );
         };
     }
-})
+});
 
 export default Layout;
